+++ src/Components/Navbar/Navbar.js
@@ -5,8 +5,7 @@ import styles from "../Navbar/Navbar.module.scss";
 
 // ? All Component and View imports
 import LogIn from "../../Components/LogIn/LogIn";
-import SignUp from "../../Components/SignUp/Signup";
-// import LogIn2 from "../test/LogIn2/LogIn2";
+// import SignUp from "../../Components/SignUp/Signup";
 
 const Navbar = () => {
     const [showLogIn, setShowLogIn] = useState(false);
@@ -22,34 +21,6 @@ const Navbar = () => {
         console.log("background was clicked close log in");
     };
 
-    const [showSignUp, setShowSignUp] = useState(false);
-
-    const openSignUp = (event) => {
-        event.preventDefault();
-        setShowSignUp(true);
-    };
-
-    const closeSignUp = (event) => {
-        event.preventDefault();
-        setShowSignUp(false);
-        console.log("background was clicked close sign up");
-    };
-
-    // Login Modal is not visible
-    if (!props.showLogin) {
-        return null;
-    }
-
-    const stop = (event) => {
-        event.stopPropagation();
-    };
-
-    // !! ASK GROUP, go to line
-    const switchToSignup = (event) => {
-        props.setShowLogIn(false);
-        setShowSignUp(true);
-    };
-
     return (
         <>
             <div className={styles["navbar-container"]}>
@@ -58,7 +29,7 @@ const Navbar = () => {
                     showLogin={showLogIn}
                     // setShowLogIn={setShowLogIn}
                 />
-                <SignUp onCloseSignup={closeSignUp} showSignup={showSignUp} />
+                {/* <SignUp onCloseSignup={closeSignUp} showSignup={showSignUp} /> */}
                 <div className={styles["icon-container"]}>
                     <div className={styles.icon}>
                         <i class="fas fa-caravan"></i>
